+++ db4e/Modules/MoneroD.py
@@ -12,10 +12,10 @@ Everything Monero Daemon
 
 from db4e.Modules.SoftwareSystem import SoftwareSystem
 from db4e.Modules.Components import (
-    ConfigFile, DataDir, InPeers, Instance, Local, LogLevel, LogName, 
-    MaxLogFiles, MaxLogSize, OutPeers, P2PBindPort, ObjectId)
+    ConfigFile, DataDir, InPeers, Instance, Local, LogLevel, LogFile, 
+    MaxLogFiles, MaxLogSize, OutPeers, P2PBindPort)
 from db4e.Constants.Fields import (MONEROD_FIELD, CONFIG_FILE_FIELD, DATA_DIR_FIELD,
-    IN_PEERS_FIELD, INSTANCE_FIELD, REMOTE_FIELD, LOG_LEVEL_FIELD, LOG_NAME_FIELD,
+    IN_PEERS_FIELD, INSTANCE_FIELD, REMOTE_FIELD, LOG_LEVEL_FIELD, LOG_FILE_FIELD,
     MAX_LOG_FILES_FIELD, MAX_LOG_SIZE_FIELD, OUT_PEERS_FIELD, P2P_BIND_PORT_FIELD,
     OBJECT_ID_FIELD)
 from db4e.Constants.Labels import (MONEROD_LABEL)
@@ -33,7 +33,7 @@ class MoneroD(SoftwareSystem):
         self.add_component(IN_PEERS_FIELD, InPeers())
         self.add_component(INSTANCE_FIELD, Instance())
         self.add_component(LOG_LEVEL_FIELD, LogLevel())
-        self.add_component(LOG_NAME_FIELD, LogName())
+        self.add_component(LOG_FILE_FIELD, LogFile())
         self.add_component(MAX_LOG_FILES_FIELD, MaxLogFiles())
         self.add_component(MAX_LOG_SIZE_FIELD, MaxLogSize())
         self.add_component(OUT_PEERS_FIELD, OutPeers())
@@ -45,7 +45,7 @@ class MoneroD(SoftwareSystem):
         self.in_peers = self.components[IN_PEERS_FIELD]
         self.instance = self.components[INSTANCE_FIELD]
         self.log_level = self.components[LOG_LEVEL_FIELD]
-        self.log_name = self.components[LOG_NAME_FIELD]
+        self.log_file = self.components[LOG_FILE_FIELD]
         self.max_log_files = self.components[MAX_LOG_FILES_FIELD]
         self.max_log_size = self.components[MAX_LOG_SIZE_FIELD]
         self.out_peers = self.components[OUT_PEERS_FIELD]